+++ common/src/main/java/com/faboslav/structurify/common/config/client/api/controller/DualController.java
@@ -3,7 +3,6 @@ package com.faboslav.structurify.common.config.client.api.controller;
 import com.faboslav.structurify.common.config.client.api.controller.element.DualControllerElement;
 import com.faboslav.structurify.common.config.client.api.option.OptionPair;
 import dev.isxander.yacl3.api.Controller;
-import dev.isxander.yacl3.api.LabelOption;
 import dev.isxander.yacl3.api.Option;
 import dev.isxander.yacl3.api.utils.Dimension;
 import dev.isxander.yacl3.gui.AbstractWidget;
@@ -11,7 +10,7 @@ import dev.isxander.yacl3.gui.TextScaledButtonWidget;
 import dev.isxander.yacl3.gui.YACLScreen;
 import net.minecraft.network.chat.Component;
 
-public record DualController<K extends Option<?>, V extends Option<?>>(LabelOption labelOption, OptionPair<K, V> optionPair) implements Controller<OptionPair<K, V>>
+public record DualController<K extends Option<?>, V extends Option<?>>(OptionPair<K, V> optionPair) implements Controller<OptionPair<K, V>>
 {
 	@Override
 	public Option<OptionPair<K, V>> option() {
@@ -28,7 +27,6 @@ public record DualController<K extends Option<?>, V extends Option<?>>(LabelOpti
 		Dimension<Integer> firstWidgetDimension = widgetDimension.withWidth((int) ((double) widgetDimension.width() * 0.5));
 		Dimension<Integer> secondWidgetDimension = widgetDimension.moved(firstWidgetDimension.width(), 0).withWidth((int) ((double) widgetDimension.width() - firstWidgetDimension.width()));
 
-		AbstractWidget labelWidget = this.labelOption.controller().provideWidget(screen, firstWidgetDimension);
 		AbstractWidget firstOptionWidget = this.optionPair.firstOption().controller().provideWidget(screen, firstWidgetDimension);
 		AbstractWidget secondOptionWidget = this.optionPair.secondOption().controller().provideWidget(screen, secondWidgetDimension);
 		TextScaledButtonWidget resetButtonWidget;
@@ -48,6 +46,6 @@ public record DualController<K extends Option<?>, V extends Option<?>>(LabelOpti
 			resetButtonWidget = null;
 		}
 
-		return new DualControllerElement(widgetDimension, labelWidget, firstOptionWidget, secondOptionWidget, resetButtonWidget);
+		return new DualControllerElement(widgetDimension, firstOptionWidget, secondOptionWidget, resetButtonWidget);
 	}
 }